+++ .iex.exs
@@ -1,5 +1,15 @@
 make_server = fn ->
-  {:ok, pid} = Spear.Connection.start_link(connection_string: "esdb://localhost:2113")
+  params = [
+    connection_string: "esdb://localhost:2113?tls=true",
+    credentials: {"admin", "changeit"},
+    opts: [
+      transport_opts: [
+        cacertfile: Path.join([__DIR__, "certs", "ca", "ca.crt"])
+      ]
+    ]
+  ]
+
+  {:ok, pid} = Spear.Connection.start_link(params)
 
   pid
 end