+++ test/spear/acl_test.exs
@@ -1,0 +1,4 @@
+defmodule Spear.AclTest do
+  use ExUnit.Case, async: true
+  doctest Spear.Acl
+end