+++ CHANGELOG.md
@@ -18,6 +18,20 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 -->
 
+## 0.1.3 - 2021-04-15
+
+### Added
+
+- Added documentation and functionality for using TLS certificates
+    - see `Spear.Connection` and the [security guide](guides/security.md)
+- Added documentation and functionality for setting the global stream ACL
+    - see `Spear.set_global_acl/4` and the `Spear.Acl` module
+- Added functionality for getting and setting stream-level metadata.
+    - `Spear.meta_stream/1`
+    - `Spear.get_stream_metadata/3`
+    - `Spear.set_stream_metadata/3`
+    - `Spear.StreamMetadata`
+
 ## 0.1.2 - 2021-04-14
 
 ### Added